+++ src/inserter.rs
@@ -1,41 +1,51 @@
-use std::{future::Future, mem};
+use std::mem;
 
 use serde::Serialize;
-use tokio::time::{Duration, Instant};
+use tokio::time::Duration;
 
 use crate::{error::Result, insert::Insert, row::Row, ticks::Ticks, Client};
 
-const DEFAULT_MAX_ENTRIES: u64 = 500_000;
+const DEFAULT_MAX_ROWS: u64 = 500_000;
 
 /// Performs multiple consecutive `INSERT`s.
 ///
+/// By default, it ends the current active `INSERT` every 500_000 rows.
+/// Use `with_max_bytes`,`with_max_rows` and `with_period` to modify this behaviour.
+///
 /// Rows are being sent progressively to spread network load.
+///
+/// All rows written by [`Inserter::write()`] between [`Inserter::commit()`] calls
+/// are sent in one `INSERT` statement.
 #[must_use]
 pub struct Inserter<T> {
     client: Client,
     table: String,
-    max_entries: u64,
+    max_bytes: u64,
+    max_rows: u64,
     send_timeout: Option<Duration>,
     end_timeout: Option<Duration>,
-    insert: Insert<T>,
+    insert: Option<Insert<T>>,
     ticks: Ticks,
-    committed: Quantities,
-    uncommitted_entries: u64,
+    pending: Quantities,
+    in_transaction: bool,
 }
 
-/// Statistics about inserted rows.
+/// Statistics about pending or inserted data.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Quantities {
-    /// How many rows ([`Inserter::write`]) have been inserted.
-    pub entries: u64,
-    /// How many nonempty transactions ([`Inserter::commit`]) have been inserted.
+    /// The number of uncompressed bytes.
+    pub bytes: u64,
+    /// The number for rows (calls of [`Inserter::write`]).
+    pub rows: u64,
+    /// The number of nonempty transactions (calls of [`Inserter::commit`]).
     pub transactions: u64,
 }
 
 impl Quantities {
     /// Just zero quantities, nothing special.
     pub const ZERO: Quantities = Quantities {
-        entries: 0,
+        bytes: 0,
+        rows: 0,
         transactions: 0,
     };
 }
@@ -48,17 +58,18 @@ where
         Ok(Self {
             client: client.clone(),
             table: table.into(),
-            max_entries: DEFAULT_MAX_ENTRIES,
+            max_bytes: u64::MAX,
+            max_rows: DEFAULT_MAX_ROWS,
             send_timeout: None,
             end_timeout: None,
-            insert: client.insert(table)?,
+            insert: None,
             ticks: Ticks::default(),
-            committed: Quantities::ZERO,
-            uncommitted_entries: 0,
+            pending: Quantities::ZERO,
+            in_transaction: false,
         })
     }
 
-    /// See [`Insert::with_max_entries()`].
+    /// See [`Insert::with_timeouts()`].
     ///
     /// Note that [`Inserter::commit()`] can call [`Insert::end()`] inside,
     /// so `end_timeout` is also applied to `commit()` method.
@@ -71,14 +82,25 @@ where
         self
     }
 
+    /// The maximum number of uncompressed bytes in one `INSERT` statement.
+    ///
+    /// Note: ClickHouse inserts batches atomically only if all rows fit in the same partition
+    /// and their number is less [`max_insert_block_size`](https://clickhouse.tech/docs/en/operations/settings/settings/#settings-max_insert_block_size).
+    ///
+    /// Unlimited by default.
+    pub fn with_max_bytes(mut self, threshold: u64) -> Self {
+        self.set_max_bytes(threshold);
+        self
+    }
+
     /// The maximum number of rows in one `INSERT` statement.
     ///
     /// Note: ClickHouse inserts batches atomically only if all rows fit in the same partition
     /// and their number is less [`max_insert_block_size`](https://clickhouse.tech/docs/en/operations/settings/settings/#settings-max_insert_block_size).
     ///
     /// `500_000` by default.
-    pub fn with_max_entries(mut self, threshold: u64) -> Self {
-        self.set_max_entries(threshold);
+    pub fn with_max_rows(mut self, threshold: u64) -> Self {
+        self.set_max_rows(threshold);
         self
     }
 
@@ -89,6 +111,12 @@ where
     /// However, it's possible to use [`Inserter::time_left()`] and set a timer up
     /// to call [`Inserter::commit()`] to check passed time again.
     ///
+    /// Extra ticks are skipped if the previous `INSERT` is still in progress:
+    /// ```text
+    /// Expected ticks: |     1     |     2     |     3     |     4     |     5     |     6     |
+    /// Actual ticks:   | work -----|          delay          | work ---| work -----| work -----|
+    /// ```
+    ///
     /// `None` by default.
     pub fn with_period(mut self, period: Option<Duration>) -> Self {
         self.set_period(period);
@@ -96,9 +124,11 @@ where
     }
 
     /// Adds a bias to the period. The actual period will be in the following range:
-    /// ```ignore
+    /// ```text
     ///   [period * (1 - bias), period * (1 + bias)]
     /// ```
+    ///
+    /// The `bias` parameter is clamped to the range `[0, 1]`.
     ///
     /// It helps to avoid producing a lot of `INSERT`s at the same time by multiple inserters.
     pub fn with_period_bias(mut self, bias: f64) -> Self {
@@ -106,22 +136,23 @@ where
         self
     }
 
-    #[deprecated(note = "use `with_period()` instead")]
-    pub fn with_max_duration(mut self, threshold: Duration) -> Self {
-        self.set_period(Some(threshold));
-        self
-    }
-
     /// See [`Inserter::with_timeouts()`].
     pub fn set_timeouts(&mut self, send_timeout: Option<Duration>, end_timeout: Option<Duration>) {
         self.send_timeout = send_timeout;
         self.end_timeout = end_timeout;
-        self.insert.set_timeouts(send_timeout, end_timeout);
+        if let Some(insert) = &mut self.insert {
+            insert.set_timeouts(self.send_timeout, self.end_timeout);
+        }
     }
 
-    /// See [`Inserter::with_max_entries()`].
-    pub fn set_max_entries(&mut self, threshold: u64) {
-        self.max_entries = threshold;
+    /// See [`Inserter::with_max_bytes()`].
+    pub fn set_max_bytes(&mut self, threshold: u64) {
+        self.max_bytes = threshold;
+    }
+
+    /// See [`Inserter::with_max_rows()`].
+    pub fn set_max_rows(&mut self, threshold: u64) {
+        self.max_rows = threshold;
     }
 
     /// See [`Inserter::with_period()`].
@@ -136,47 +167,58 @@ where
         self.ticks.reschedule();
     }
 
-    #[deprecated(note = "use `set_period()` instead")]
-    pub fn set_max_duration(&mut self, threshold: Duration) {
-        self.ticks.set_period(Some(threshold));
-    }
-
     /// How much time we have until the next tick.
     ///
     /// `None` if the period isn't configured.
     pub fn time_left(&mut self) -> Option<Duration> {
-        Some(
-            self.ticks
-                .next_at()?
-                .saturating_duration_since(Instant::now()),
-        )
+        self.ticks.time_left()
     }
 
-    /// Serializes and writes to the socket a provided row.
+    /// Returns statistics about data not yet inserted into ClickHouse.
+    pub fn pending(&self) -> &Quantities {
+        &self.pending
+    }
+
+    /// Serializes the provided row into an internal buffer.
+    ///
+    /// To check limits and sent to ClickHouse, call [`Inserter::commit()`].
     ///
     /// # Panics
     /// If called after previous call returned an error.
     #[inline]
-    pub fn write<'a>(&'a mut self, row: &T) -> impl Future<Output = Result<()>> + 'a + Send
+    pub fn write(&mut self, row: &T) -> Result<()>
     where
         T: Serialize,
     {
-        self.uncommitted_entries += 1;
-        self.insert.write(row)
-    }
-
-    /// Checks limits and ends a current `INSERT` if they are reached.
-    pub async fn commit(&mut self) -> Result<Quantities> {
-        if self.uncommitted_entries > 0 {
-            self.committed.entries += self.uncommitted_entries;
-            self.committed.transactions += 1;
-            self.uncommitted_entries = 0;
+        if self.insert.is_none() {
+            self.init_insert()?;
         }
 
-        let now = Instant::now();
+        match self.insert.as_mut().unwrap().do_write(row) {
+            Ok(bytes) => {
+                self.pending.bytes += bytes as u64;
+                self.pending.rows += 1;
 
-        Ok(if self.is_threshold_reached(now) {
-            let quantities = mem::replace(&mut self.committed, Quantities::ZERO);
+                if !self.in_transaction {
+                    self.pending.transactions += 1;
+                    self.in_transaction = true;
+                }
+
+                Ok(())
+            }
+            Err(err) => {
+                self.pending = Quantities::ZERO;
+                Err(err)
+            }
+        }
+    }
+
+    /// Checks limits and ends the current `INSERT` if they are reached.
+    pub async fn commit(&mut self) -> Result<Quantities> {
+        self.in_transaction = false;
+
+        Ok(if self.limits_reached() {
+            let quantities = mem::replace(&mut self.pending, Quantities::ZERO);
             let result = self.insert().await;
             self.ticks.reschedule();
             result?;
@@ -186,23 +228,36 @@ where
         })
     }
 
-    /// Ends a current `INSERT` and whole `Inserter` unconditionally.
+    /// Ends the current `INSERT` and whole `Inserter` unconditionally.
     ///
     /// If it isn't called, the current `INSERT` is aborted.
-    pub async fn end(self) -> Result<Quantities> {
-        self.insert.end().await?;
-        Ok(self.committed)
+    pub async fn end(mut self) -> Result<Quantities> {
+        self.insert().await?;
+        Ok(self.pending)
     }
 
-    fn is_threshold_reached(&self, now: Instant) -> bool {
-        self.committed.entries >= self.max_entries
-            || self.ticks.next_at().map_or(false, |next_at| now >= next_at)
+    fn limits_reached(&self) -> bool {
+        self.pending.rows >= self.max_rows
+            || self.pending.bytes >= self.max_bytes
+            || self.ticks.reached()
     }
 
     async fn insert(&mut self) -> Result<()> {
-        let mut new_insert = self.client.insert(&self.table)?; // Actually it mustn't fail.
+        if let Some(insert) = self.insert.take() {
+            insert.end().await?;
+        }
+        Ok(())
+    }
+
+    #[cold]
+    #[inline(never)]
+    fn init_insert(&mut self) -> Result<()> {
+        debug_assert!(self.insert.is_none());
+        debug_assert_eq!(self.pending, Quantities::ZERO);
+
+        let mut new_insert: Insert<T> = self.client.insert(&self.table)?;
         new_insert.set_timeouts(self.send_timeout, self.end_timeout);
-        let insert = mem::replace(&mut self.insert, new_insert);
-        insert.end().await
+        self.insert = Some(new_insert);
+        Ok(())
     }
 }