+++ src/row.rs
@@ -4,6 +4,7 @@ pub trait Row {
     const COLUMN_NAMES: &'static [&'static str];
 
     // TODO: count
+    // TODO: different list for SELECT/INSERT (de/ser)
 }
 
 // Actually, it's not public now.
@@ -132,6 +133,21 @@ mod tests {
         struct TopLevel {
             one: u32,
             #[serde(skip_serializing)]
+            two: u32,
+        }
+
+        assert_eq!(join_column_names::<TopLevel>().unwrap(), "`one`");
+    }
+
+    #[test]
+    fn it_skips_deserializing() {
+        use serde::Deserialize;
+
+        #[derive(Row, Deserialize)]
+        #[allow(dead_code)]
+        struct TopLevel {
+            one: u32,
+            #[serde(skip_deserializing)]
             two: u32,
         }