+++ tests/it/compression.rs
@@ -2,8 +2,6 @@ use serde::{Deserialize, Serialize};
 
 use clickhouse::{Client, Compression, Row};
 
-mod common;
-
 async fn check(client: Client) {
     #[derive(Debug, Row, Serialize, Deserialize)]
     struct MyRow<'a> {
@@ -23,14 +21,11 @@ async fn check(client: Client) {
         .await
         .unwrap();
 
-    let mut inserter = client.inserter("test").unwrap();
-
+    let mut insert = client.insert("test").unwrap();
     for i in 0..200_000 {
-        inserter.write(&MyRow { no: i, name: "foo" }).await.unwrap();
-        inserter.commit().await.unwrap();
+        insert.write(&MyRow { no: i, name: "foo" }).await.unwrap();
     }
-
-    inserter.end().await.unwrap();
+    insert.end().await.unwrap();
 
     // Check data.
 
@@ -44,25 +39,22 @@ async fn check(client: Client) {
     assert_eq!(sum_len, 600_000);
 }
 
-#[common::named]
 #[tokio::test]
 async fn none() {
-    let client = common::prepare_database!().with_compression(Compression::None);
+    let client = prepare_database!().with_compression(Compression::None);
     check(client).await;
 }
 
 #[cfg(feature = "lz4")]
-#[common::named]
 #[tokio::test]
 async fn lz4() {
-    let client = common::prepare_database!().with_compression(Compression::Lz4);
+    let client = prepare_database!().with_compression(Compression::Lz4);
     check(client).await;
 }
 
 #[cfg(feature = "lz4")]
-#[common::named]
 #[tokio::test]
 async fn lz4_hc() {
-    let client = common::prepare_database!().with_compression(Compression::Lz4Hc(4));
+    let client = prepare_database!().with_compression(Compression::Lz4Hc(4));
     check(client).await;
 }