+++ tests/it/query.rs
@@ -2,12 +2,9 @@ use serde::{Deserialize, Serialize};
 
 use clickhouse::{error::Error, AsyncInsertOptions, Row};
 
-mod common;
-
-#[common::named]
 #[tokio::test]
 async fn smoke() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     #[derive(Debug, Row, Serialize, Deserialize)]
     struct MyRow<'a> {
@@ -54,10 +51,9 @@ async fn smoke() {
     }
 }
 
-#[common::named]
 #[tokio::test]
 async fn async_insert() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     #[derive(Debug, Row, Serialize, Deserialize)]
     struct MyRow<'a> {
@@ -127,10 +123,9 @@ async fn async_insert() {
     }
 }
 
-#[common::named]
 #[tokio::test]
 async fn fetch_one_and_optional() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     client
         .query("CREATE TABLE test(n String) ENGINE = MergeTree ORDER BY n")
@@ -163,10 +158,9 @@ async fn fetch_one_and_optional() {
 }
 
 // See #19.
-#[common::named]
 #[tokio::test]
 async fn long_query() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     client
         .query("CREATE TABLE test(n String) ENGINE = MergeTree ORDER BY n")
@@ -187,10 +181,9 @@ async fn long_query() {
 }
 
 // See #22.
-#[common::named]
 #[tokio::test]
 async fn big_borrowed_str() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     #[derive(Debug, Row, Serialize, Deserialize)]
     struct MyRow<'a> {
@@ -226,10 +219,9 @@ async fn big_borrowed_str() {
 }
 
 // See #31.
-#[common::named]
 #[tokio::test]
 async fn all_floats() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     client
         .query("CREATE TABLE test(no UInt32, f Float64) ENGINE = MergeTree ORDER BY no")