+++ tests/it/uuid.rs
@@ -5,12 +5,9 @@ use uuid::Uuid;
 
 use clickhouse::Row;
 
-mod common;
-
-#[common::named]
 #[tokio::test]
 async fn smoke() {
-    let client = common::prepare_database!();
+    let client = prepare_database!();
 
     #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Row)]
     struct MyRow {